+++ Frends.FixedWidthFlatFile.ConvertToXML/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.0.0] - 2026-09-10
+### Changed
+- Target framework changed to net8.0.
+- Added `Options` parameter with `ThrowErrorOnFailure` and `ErrorMessageOnFailure` properties, allowing errors to be returned as a failed result instead of throwing an exception.
+- `Result` object now includes `Success` and `Error` properties in addition to `Data`.
+- Renamed the `data` parameter to `input`.
+
 ## [1.1.0] - 2026-04-16
 ### Updated
 - Updated Result.Data documentation and task description.